refactor(LottoValidator): 디렉토리를 이동한다
